DataTable: Fixed Length property being a decimal
I can't remember why it was a decimal, but that meant, that many
conversions would fail.
A long should be fine for counting the elements.

+++ CalcLang/Interpreter/DataTable.cs
@@ -74,9 +74,9 @@ namespace CalcLang.Interpreter
             [GRE] = ExpressionType.GreaterThan
         };
 
-        public decimal Length => length;
+        public long Length => length;
 
-        private decimal length;
+        private long length;
         private DataTable keys;
 
         private bool invalidated;
